backend/requirements_relaxed.txt
# FastAPI and ASGI Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
authlib>=1.3.0

# Cache and Queue
redis>=5.0.1
fastapi-cache2[redis]>=0.2.1
celery>=5.3.6

# Storage
boto3>=1.34.28

# Search
elasticsearch>=8.12.0

# Utilities
python-dotenv>=1.0.0
email-validator>=2.1.0
httpx>=0.26.0
